Fix int32 overflow in substruct pair indexing
For batches large enough that numTargets * numQueries exceeds INT32_MAX
(e.g. 10M targets x 480 queries = 4.8B pairs), the flat pair index
computed as `targetIdx * numQueries + queryIdx` wrapped to a negative
int. That index was then sign-extended to size_t when used to subscript
HasSubstructMatchResults::hasMatch and the counts vector, causing
out-of-bounds writes and segfaults in hasSubstructMatch and
countSubstructMatches. The same overflow corrupted unordered_map keys
in SubstructSearchResults::matches.

Widen pairIndex computations and the map key type to int64_t. Cast to
size_t at indexing sites. Kernel-side pair indices are unchanged because
they operate in batch-local target coordinates bounded by batchSize.

+++ src/substruct/substruct_search_internal.cpp
@@ -155,8 +155,8 @@ void processWithRDKitFallback(const RDKit::ROMol*       target,
   if (boolResults) {
     boolResults->setMatch(targetIdx, queryIdx, true);
   } else if (countResults) {
-    const int pairIdx        = targetIdx * results.numQueries + queryIdx;
-    (*countResults)[pairIdx] = matchCount;
+    const int64_t pairIdx                         = static_cast<int64_t>(targetIdx) * results.numQueries + queryIdx;
+    (*countResults)[static_cast<size_t>(pairIdx)] = matchCount;
   } else {
     std::vector<std::vector<int>> convertedMatches;
     convertedMatches.reserve(rdkitMatches.size());
@@ -399,9 +399,9 @@ void accumulateMiniBatchResultsCounts(GpuExecutor&               executor,
   std::lock_guard<std::mutex> lock(resultsMutex);
 
   for (int i = 0; i < executor.plan.numPairsInMiniBatch; ++i) {
-    const auto [targetIdx, queryIdx] = resolvePairIndices(i, ctx, hostBuffer);
-    const int globalPairIdx          = targetIdx * ctx.numQueries + queryIdx;
-    counts[globalPairIdx]            = hostBuffer.matchCounts[i];
+    const auto [targetIdx, queryIdx]           = resolvePairIndices(i, ctx, hostBuffer);
+    const int64_t globalPairIdx                = static_cast<int64_t>(targetIdx) * ctx.numQueries + queryIdx;
+    counts[static_cast<size_t>(globalPairIdx)] = hostBuffer.matchCounts[i];
   }
 }
